Extract GallerySelectionBadgeView
